+++ hibernate-core/src/main/java/org/hibernate/engine/query/spi/NativeSQLQueryPlan.java
@@ -87,7 +87,7 @@ public class NativeSQLQueryPlan implements Serializable {
 					customQuery.getSQL() );
 		}
 		if ( loc instanceof Integer ) {
-			return new int[] { ((Integer) loc ).intValue() };
+			return new int[] { (Integer) loc };
 		}
 		else {
 			return ArrayHelper.toIntArray( (List) loc );
