+++ hibernate-core/src/main/java/org/hibernate/hql/internal/ast/exec/BasicExecutor.java
@@ -96,7 +96,7 @@ public class BasicExecutor implements StatementExecutor {
 				}
 				if ( selection != null ) {
 					if ( selection.getTimeout() != null ) {
-						st.setQueryTimeout( selection.getTimeout().intValue() );
+						st.setQueryTimeout( selection.getTimeout() );
 					}
 				}
